[#7682] leave reportingUser as it is
